♻️ Moved authentication to auth model
