+++ server/authentication/models.py
@@ -1,7 +1,13 @@
-import pymongo
 import os
+from typing import Dict, Union
+
+import pymongo
 from free_watch import conf
-from free_watch.errorfactory import InvalidUid
+from free_watch.errorfactory import (
+    ExcessiveUnitsError,
+    InvalidCredentialsError,
+    InvalidUid,
+)
 
 
 class AuthModel:
@@ -23,3 +29,44 @@ class AuthModel:
     def id_from_user(self, user_name: str):
         user = self.db.users.find_one({"user_name": user_name})
         return user["unit_id"]
+
+    def register_user(self, doc: Dict[str, Union[str, int]], limit: int = 50):
+        """Register new users and assign units
+
+        Args:
+            doc (Dict[str, Union[str, int]]): user data
+        """
+
+        _units = doc.get("units")
+        if _units > limit:
+            raise ExcessiveUnitsError(
+                detail={
+                    "error": f"Excessive no. of units {_units} current max units are {self.max_entry}"
+                }
+            )
+
+        doc = {**{"unit_id": self.get_uid(length=16)}, **doc}
+        self.db.users.insert_one(doc)
+        self.db.units.insert_one({"unit_id": doc["unit_id"], "data": []})
+
+    def credentials(self, password, email):
+        user = self.db.users.find_one({"password": password, "email": email})
+        if user:
+            return user
+        raise InvalidCredentialsError(detail={"error": "Invalid credentials"})
+
+    def reset_password(self, old_passwd: str, email_id: str, new_passwd: str):
+        """Takes in hashed passwords updates
+           password if user found.
+
+        Args:
+            old_pswd (str): old hashed password
+            new_pswd (str): new hashed password
+            email_id (str): email_id
+        """
+
+        self.db.users.find_one_and_update(
+            {"password": old_passwd, "email": email_id},
+            update={"$set": {"password": new_passwd}},
+        )
+        return None
