+++ server/free_watch/settings.py
@@ -50,6 +50,14 @@ MIDDLEWARE = [
     "authentication.middleware.AuthMiddleWare",
 ]
 
+AUTH_PASSWORD_VALIDATORS = [
+    "django.contrib.auth.password_validation.MinimumLengthValidator",
+    "django.contrib.auth.password_validation.UserAttributeSimilarityValidator",
+    "django.contrib.auth.password_validation.CommonPasswordValidator",
+    "django.contrib.auth.password_validation.NumericPasswordValidator",
+]
+
+
 ROOT_URLCONF = "free_watch.urls"
 
 TEMPLATES = [
